+++ how-to/customize-workspace/client/src/modules/init-options/interop/index.ts
@@ -1,5 +1,5 @@
 import type { ModuleImplementation, ModuleTypes } from "../../../module-shapes";
-import * as initOptionsImplementation from "./action-handlers";
+import * as initOptionsImplementation from "./action-handler";
 
 export const entryPoints: { [type in ModuleTypes]?: ModuleImplementation } = {
 	initOptions: initOptionsImplementation
